refactor(LanczSolver): Refactor to suit gqten v0.2

+++ include/gqmps2/algorithm/lanczos_solver.h
@@ -3,7 +3,7 @@
 /*
 * Author: Rongyang Sun <sun-rongyang@outlook.com>
 * Creation Date: 2020-08-08 14:22
-* 
+*
 * Description: GraceQ/MPS2 project. Lanczos solver.
 */
 
@@ -13,6 +13,9 @@
 */
 #ifndef GQMPS2_ALGORITHM_LANCZOS_SOLVER_H
 #define GQMPS2_ALGORITHM_LANCZOS_SOLVER_H
+
+
+#include <stdlib.h>     // size_t
 
 
 namespace gqmps2 {
@@ -28,17 +31,17 @@ struct LanczosParams {
   @param error The Lanczos tolerated error.
   @param max_iterations The maximal iteration times.
   */
-  LanczosParams(double err, long max_iter) :
+  LanczosParams(double err, size_t max_iter) :
       error(err), max_iterations(max_iter) {}
   LanczosParams(double err) : LanczosParams(err, 200) {}
   LanczosParams(void) : LanczosParams(1.0E-7, 200) {}
   LanczosParams(const LanczosParams &lancz_params) :
       LanczosParams(lancz_params.error, lancz_params.max_iterations) {}
 
-  double error;           ///< The Lanczos tolerated error.
-  long max_iterations;    ///< The maximal iteration times.
+  double error;             ///< The Lanczos tolerated error.
+  size_t max_iterations;    ///< The maximal iteration times.
 };
-} /* gqmps2 */ 
+} /* gqmps2 */
 
 
 // Implementation details
